Use CLUSTER_ID as a fallback for CASTAI_CLUSTER_ID
