+++ cmd/imagescan/config/config.go
@@ -1,9 +1,7 @@
 package config
 
 import (
-	"errors"
 	"io/fs"
-	"os"
 	"time"
 
 	"github.com/kelseyhightower/envconfig"
@@ -33,11 +31,7 @@ const (
 )
 
 type Config struct {
-	CastaiClusterID   string `envconfig:"CASTAI_CLUSTER_ID" required:"true"`
-	CastaiAPIGrpcAddr string `envconfig:"CASTAI_API_GRPC_ADDR" required:"true"`
-	// The api key is required, but we support two differnt ways of setting it.
-	CastaiAPIKey       string `envconfig:"CASTAI_API_KEY"`
-	CastaiGRPCInsecure bool   `envconfig:"CASTAI_GRPC_INSECURE"`
+	CastaiGRPCInsecure bool `envconfig:"CASTAI_GRPC_INSECURE"`
 
 	BlobsCacheURL     string        `envconfig:"COLLECTOR_BLOBS_CACHE_URL"`
 	ImageID           string        `envconfig:"COLLECTOR_IMAGE_ID" required:"true"`
@@ -61,15 +55,6 @@ func FromEnv() (Config, error) {
 	var cfg Config
 	if err := envconfig.Process("", &cfg); err != nil {
 		return Config{}, err
-	}
-
-	if cfg.CastaiAPIKey == "" {
-		// fall back to non `CASTAI` prefix env variable
-		if apiKey, found := os.LookupEnv("API_KEY"); found {
-			cfg.CastaiAPIKey = apiKey
-		} else {
-			return Config{}, errors.New("required environment variable not set: CASTAI_API_KEY or API_KEY are missing")
-		}
 	}
 
 	return cfg, nil
